ds_backend-main/requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==2.2.8
django-cors-headers==3.10.0
django-countries==7.3.2
django-rest-knox==4.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.8.6
pycparser==2.21
python-slugify==6.1.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.7
uWSGI==2.0.18
moviepy==1.0.3